+++ src/components/index.ts
@@ -71,7 +71,7 @@ export { ClipboardCopy } from './patternfly-wrappers/clipboard-copy';
 export { FileUpload } from './patternfly-wrappers/file-upload';
 export { LabelGroup } from './patternfly-wrappers/label-group';
 export { LoginForm } from './patternfly-wrappers/login-form';
-export { HubPagination } from './patternfly-wrappers/pagination';
+export { Pagination } from './patternfly-wrappers/pagination';
 export { Tooltip } from './patternfly-wrappers/tooltip';
 export { AccessTab } from './rbac/access-tab';
 export { DeleteGroupModal } from './rbac/delete-group-modal';
@@ -119,6 +119,7 @@ export { Details } from './shared/details';
 export { DownloadCount } from './shared/download-count';
 export { ExternalLink } from './shared/external-link';
 export { HubListToolbar } from './shared/hub-list-toolbar';
+export { HubPagination } from './shared/hub-pagination';
 export { LanguageSwitcher } from './shared/language-switcher';
 export { LinkTabs, LinkTabsProps } from './shared/link-tabs';
 export { LoginLink } from './shared/login-link';
